[#47015] Replace Statuses::TableCell by ViewComponent
https://community.openproject.org/work_packages/47015
