+++ app/components/statuses/table_component.rb
@@ -1,0 +1,73 @@
+# frozen_string_literal: true
+
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) 2012-2023 the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+module Statuses
+  class TableComponent < ::TableComponent
+    def initial_sort
+      %i[id asc]
+    end
+
+    def sortable?
+      false
+    end
+
+    def columns
+      headers.map(&:first)
+    end
+
+    def inline_create_link
+      link_to new_status_path,
+              aria: { label: t(:label_work_package_status_new) },
+              class: 'wp-inline-create--add-link',
+              title: t(:label_work_package_status_new) do
+        op_icon('icon icon-add')
+      end
+    end
+
+    def empty_row_message
+      I18n.t :no_results_title_text
+    end
+
+    def headers
+      [
+        [:name, { caption: Status.human_attribute_name(:name) }],
+        [:color, { caption: Status.human_attribute_name(:color) }],
+        [:default?, { caption: Status.human_attribute_name(:is_default) }],
+        [:closed?, { caption: Status.human_attribute_name(:is_closed) }],
+        [:readonly?, { caption: Status.human_attribute_name(:is_readonly) }],
+        [:sort, { caption: I18n.t(:label_sort) }]
+      ].tap do |default|
+        if WorkPackage.use_status_for_done_ratio?
+          default.insert 2, [:done_ratio, { caption: WorkPackage.human_attribute_name(:done_ratio) }]
+        end
+      end
+    end
+  end
+end
